test: move flask base to helper
avoid duplicates
