+++ demos/jans-tent/clientapp/utils/dcr_from_config.py
@@ -6,6 +6,7 @@ import json
 OP_URL = cfg.ISSUER
 REDIRECT_URIS = cfg.REDIRECT_URIS
 
+
 def setup_logging() -> None:
     logging.getLogger('oic')
     logging.getLogger('urllib3')
@@ -17,10 +18,9 @@ def setup_logging() -> None:
 
 def register() -> None:
     logger = logging.getLogger(__name__)
-    logger.info('Trying to auto-register client from configuration')
     client_handler = ClientHandler(OP_URL, REDIRECT_URIS)
     json_client_info = json.dumps(client_handler.get_client_dict(), indent=4)
     with open('client_info.json', 'w') as outfile:
-        logger.info('Writing registered client information to clien_info.json')
+        logger.info('Writing registered client information to client_info.json')
         outfile.write(json_client_info)
