+++ demos/jans-tent/tests/unit_integration/test_configuration_endpoint.py
@@ -1,9 +1,8 @@
-from unittest import TestCase
 import clientapp
 from flask import Flask, url_for
 from typing import List
-import os
 import json
+from helper import FlaskBaseTestCase
 
 
 def app_endpoints(app: Flask) -> List[str]:
@@ -22,17 +21,6 @@ def valid_client_configuration():
         "client_secret": "my-client-secret",
         "op_metadata_url": "https://op.com/.well-known/openidconfiguration"
     }
-
-
-class FlaskBaseTestCase(TestCase):
-    def setUp(self):
-        self.app = clientapp.create_app()
-        self.app.testing = True
-        self.app_context = self.app.test_request_context()
-        self.app_context.push()
-        self.client = self.app.test_client()
-        #self.oauth = OAuth(self.app)
-        os.environ['AUTHLIB_INSECURE_TRANSPORT'] = "1"
 
 
 class TestConfigurationEndpoint(FlaskBaseTestCase):
